Move clipboard into its own module
